moved libraryStack folder
